Object oriented methods are extracted into 3 model files(with some new files e.g. preprocess.py), and part of the source code in labelImg.py has been simplified
