+++ libs/detector/ssd/model.py
@@ -1,58 +1,43 @@
 #!/usr/bin/python3 python
 # encoding: utf-8
 '''
-@author: 孙昊
-@contact: smartadpole@163.com
+@author: 焦子傲
+@contact: 1054568970@.com
 @file: model.py
-@time: 2021/2/2 下午5:55
-@desc: 
+@time: 2021/4/14 10:42
+@desc:
 '''
-import sys, os
+import os
+from libs.detector.ssd.onnxmodel import ONNXModel
+from libs.detector.ssd.preprocess import pre_process
+from libs.detector.ssd.postprocess.ssd import PostProcessor_SSD
+from libs.detector.ssd.postprocess.ssd import THRESHOLD
 
-CURRENT_DIR = os.path.dirname(__file__)
-sys.path.append(os.path.join(CURRENT_DIR, '../../'))
+class SSD(object):
+    def __init__(self, file='./config/cleaner/ssd.onnx'):
+        if os.path.isfile(file):
+            self.net = ONNXModel(file)
+        else:
+            raise IOError("no such file {}".format(self.model_file_4_detect))
 
-import argparse
-import numpy as np
+    def forward(self, image, class_names_4_detect):
+        image, ratio = pre_process(image)
+        out = self.net.forward(image)
+        results_batch = PostProcessor_SSD(out[0], out[1], out[2])
 
-# -*-coding: utf-8 -*-
+        # TODO : get rect
+        shapes = []
+        for result in results_batch:
+            if len(result) > 0:
+                result = result.tolist()
+                for r in result:
+                    x, y, x2, y2, label, score = r
+                    x, y, x2, y2, label = int(x) * ratio[0], int(y) * ratio[1], int(x2) * ratio[0], int(y2) * ratio[
+                        1], int(label)
+                    if score < THRESHOLD[label] or label == 0:
+                        continue
 
-import os, sys
-import onnxruntime
+                    shapes.append(
+                        (class_names_4_detect[label - 1], [(x, y), (x2, y), (x2, y2), (x, y2)], None, None, False))
 
-class ONNXModel():
-    def __init__(self, onnx_file):
-        self.onnx_session = onnxruntime.InferenceSession(onnx_file)
-        self.input_name = self.get_input_name(self.onnx_session)
-        self.output_name = self.get_output_name(self.onnx_session)
-        print("input_name:{}".format(self.input_name))
-        print("output_name:{}".format(self.output_name))
-
-    def get_output_name(self, onnx_session):
-        output_name = []
-        for node in onnx_session.get_outputs():
-            output_name.append(node.name)
-        return output_name
-
-    def get_input_name(self, onnx_session):
-        input_name = []
-        for node in onnx_session.get_inputs():
-            input_name.append(node.name)
-        return input_name
-
-    def get_input_feed(self, input_name, image:np.ndarray):
-        input_feed = {}
-        for name in input_name:
-            input_feed[name] = image
-        return input_feed
-
-    def forward(self, image:np.ndarray):
-        input_feed = self.get_input_feed(self.input_name, image)
-        scores = self.onnx_session.run(self.output_name, input_feed=input_feed)
-        return scores
-
-
-
-def to_numpy(tensor):
-    print(tensor.device)
-    return tensor.detach().cpu().numpy() if tensor.requires_grad else tensor.cpu().numpy()
+        return shapes
